+++ docker-deploy-prod.sh
@@ -53,6 +53,15 @@ then
       register_definition
       update_service  
 
+      # exercises
+      service="sbr-exercises-prod-service"
+      template="ecs_exercises_prod_taskdefinition.json"
+      task_template=$(cat "ecs/$template")
+      task_def=$(printf "$task_template" $AWS_ACCOUNT_ID $AWS_RDS_EXERCISES_URI)
+      echo "$task_def"
+      register_definition
+      update service
+
       # swagger
       service="sbr-swagger-prod-service"  
       template="ecs_swagger_prod_taskdefinition.json"