+++ docker-push.sh
@@ -11,6 +11,7 @@ then
   elif [[ "$TRAVIS_BRANCH" == "production" ]]; then
     export DOCKER_ENV=prod
     export REACT_APP_USERS_SERVICE_URL="http://www.smallbatchreact.com"
+    export REACT_APP_EXERCISES_SERVICE_URL="http://www.smallbatchreact.com"
     export DATABASE_URL="$AWS_RDS_URI" 
     export SECRET_KEY="$PRODUCTION_SECRET_KEY"  
   fi